docker/requirements.txt
numpy
opencv-python-headless
Pillow
python-telegram-bot==20.5
ultralytics
nest_asyncio